Implemented timer with dynamic update
Right now starting each timer will create a new progress array entry,
stopping the timer will generate time passed between start to finish of
this progress entry.

If you switch to other tabs when the timer is still running, and come
back, you will be unable to stop the timer “on the UI”, the progress is
still updated correctly inside the DB. Need to look into this later.